requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
streamlit==1.22.0
requests==2.28.2
python-dotenv==0.21.1
pydantic==1.10.7
itsdangerous==2.1.2